Java Programming
Create an abstract class GeometricObject which will have two variables
fillColor and borderColor. In the constructor setting the fillColor as “white”
and the borderColor as “black” as default values. Write getter methods
getFillColour() and getBorderColor() to return the values. The class should
have two abstract methods findArea() and findCircumference(). Write a
subclass of GeometricObject called Triangle to calculate
area and circumference for a triangle.